add support for detecting Go dependencies via go.sum
