+++ config.py
@@ -1,0 +1,1 @@
+GITHUB_API = "https://api.github.com"
